First attempt
Made the setup file which is pretty much copy paste the useful code given in the beggining of the lab notes in order for them to be loaded easily  with :l Lab1setup.

Also did an attempt at exercise 1 of the workshop as a test but cannot get a working code.